PYIC-3478: Add ProcessStepResponse
We want to have states in the journey map that can define some
processing to do. This adds a new step response that a state can return
that can be configured with a lambda and some input data for it. The
intention is this will be wired in to the journey engine step function
to run the defined lambda and pass it the data.

Currently this isn't wired in - we need a lambda before we can do that.