PYIC-3510: Correlate data for F2F VC
The f2f VC is collected async. We weren't doing the usual checks on the
data in contains to be sure it correlates with the data in the other
VCs. This meant data was being sent to SPoT that was being rejected.

This change will do the correlation in the check-existing-identity
lambda and send the user to the F2F fail page if the data doesn't match.
The user will then have the opportunity to reset their identity.

We don't need to take any action of the data doesn't correlate if we're
not on an F2F journey. This is an expected scenario and will be picked
up by them not matching a profile and be sent to the reset-journey
lambda separately.

This also moves the call to updateVcStatues higher in the handler so
it's available for the correlation check. This way the F2F VC is
included.