PYIC-3539: Maintain users connection with a CRI
If a user has been sent to a CRI they'll need to use the same
connections when they are returned to core. That is, they'll need to do
the OAuth dance with the same CRI they were sent to.
This is an important distinction in the case that we have rotated the
config for a CRI - perhaps new endpoint, new keys etc. If this change
happens after a user has initiated an OAuth session with a CRI but
before they return, they need to use the config at the point they
initiated it to complete the journey.

This stores the connection used in their CRI Oauth session details. That
connection is then re-used when coming back to core.

There are some places where we fetch CRI config independently of having
a users active CRI Oauth session. In the async journey for example, and
when building an identity. We do this by looking up the issuer of a VC
and then fetching config based on this. This implied that a CRI
shouldn't change their component ID.