PYIC-3539: Use active connection directly less
In a few places we were fetching a CRIs config using the method in the
config service for fetching the active connection.
In these places we were just trying to determine the CRIs component ID.
We have a method for doing that without pulling back the entire config.

It also means that we're using the "active" connection less directly.
Sure, the component ID method is using it under the hood, but we can
refactor that at some point in the future if we needed to.