PYIC-3531: Load API key by connection
The API key used to authenticate with the CRIs is stored in secrets
manager. This adds the ability to have an API key per CRI connection.
The new path mirrors how we structure the SSM config, but the value is
store in secrets manager.

The current API keys have been copied to the new paths in staging,
integration and prod. Effectively `connections/main/` has been added to
each API key, ready for when this code is used.