PYIC-3581: Include mitigation details in audit event
This adds a couple of bits of info to the `IPV_IDENTITY_ISSUED` audit
event issued in the build-user-identity handler. This idea is to allow
us to track how successful the mitigations work has been. For all
identities issued we'll be able to see if the user has a mitigation.
We'll also be able to see if they were breaching the CI threshold.