BAU: Fix same session bug with journey map
Enabling same session mitigations introduces a bug in separate and same
session mitigation.

The way we activate separate session mitigation is by updating the CiMit
config to send a user on a mitigation journey for same session, rather
than failing them.

This means that if a user is at DCMAW on a mitigation journey (same or
separate), and gets a failed response with no CIs, we attempt to send
them back on the same session mitigation journey. This is because they
still have the CI that needs mitigating and they still have a breaching
score due to no mitigation check score being in place.

Attempting to send them on this journey fails as we don't define that
event on the dcmaw mitigation states.

This scenario is the same as the thin file scenario. In DCMAW a user can
get this by failing the liveness likeness check.

This fixes the immediate bug by updating the journey map to handle the
mitigation event on the mitigation state. This is more of a band aid
than a proper fix.