BAU: Fix same session bug with last CRI check
Enabling same session mitigations introduces a bug in separate and same
session mitigation.

The way we activate separate session mitigation is by updating the CiMit
config to send a user on a mitigation journey for same session, rather
than failing them.

This means that if a user is at DCMAW on a mitigation journey (same or
separate), and gets a failed response with no CIs, we attempt to send
them back on the same session mitigation journey. This is because they
still have the CI that needs mitigating and they still have a breaching
score due to no mitigation check score being in place.

Attempting to send them on this journey fails as we don't define that
event on the dcmaw mitigation states.

This scenario is the same as the thin file scenario. In DCMAW a user can
get this by failing the liveness likeness check.

This fixes the bug by checking if the last CRI visited is in the
mitigation configuration for that particular CI. If it is, we assume
that mitigation was unsuccessful and don't attempt another mitigation
journey. If the user's session has no visited CRIs then they are
starting a fresh journey and we are doing the initial CI score check.
This means we should send them on the mitigation journey.