PROBLEM-81: Reorder checks in eval-gpg45-scores lambda
We've seen SPoT receive some iddentity bundles which it has then failed
due to name correlation issues. These identities shouldn't have made it
past core.

When we check for a matched profile in eval-gpg45-scores lamdba we set
the VOT value in the users session as P2 if we match a profile.

After we do this we were checking for name correlation issues. If the
names didn't correlate we'd return a no-match response to the user. But
the P2 was already set.

This meant that a user would be told they'd failed, then returned to
orchestrator. Their identity is then fetched and a bundle with P2 is
returned, despite failing name correlation. This was then sent to SPoT
and rejected.

This re-orders the checks in eval-gpg45-lambda to only set a P2 once the
other checks have passed. It also updates all the tests to ensure that
the correct value is set in the users session.