PROBLEM-81: Check for data correlation in check-existing-identity
We were only checking for data correlation in the
check-existing-identity lambda when a user had just come back from F2F.
In that scenario we send the user to a special page before allowing them
to reset their identity.

If they had come back from somewhere else and had uncorrelated data we
weren't checking.

If a user had all the VCs needed for a profile but had a name
correlation error - picked up from their last VC, we were issuing an
identity. This was being pricked up by SPoT.

Rather than issuing a P2 we should just be resetting the users identity.