BAU: Write the session after setting VOT
Write the session to the database after we update the VOT.

A previous update moved the order of some calls so that the session was no longer being written after setting the vot when a profile was matched. This change adds writing the session to the database after setting the vot. We should look to refactor this code to make it clearer what the sequence of events and session writing clearer to avoid similar issues in the future.