BAU: Strengthen unit tests around VOT setting
There are two places that we decided if a user has met a profile. If
they have, we set the VOT value on their session.

We've had a couple of incidents where code changes have led to the VOT
being set, but then the session not actually updated. Meaning that the
VOT was not persisted causing users journey to have an unexpected
failure in their journey.

The unit tests we had checking that the VOT level was correctly set were
not strong enough, allowing this issue to slip through the net. They
were checking that the session had the correct value, but not that the
session had been updated, or that it been updated _after_ the VOT was
set. This does that check.

It also adds a check to a number of tests for the negative. This may
seem redundant, especially where we're already checking that the session
is never updated, but I think it makes the intent clearer. And may help
to catch issues in future.