PYIC-3635: Accept legacy and new cimit config
We want to simplify the way we configure the journey events for
mitigation. This means not maintaining separate events for same and
separate session.

We need to deploy this out in a way that won’t cause downtime.

First job is to allow core to handle both types of config.

This change will attempt to use the original config style and fall back
to the new style if it can't (because the config has been updated).