BAU: Use input streams in state machine initializer
This updates the state machine initializer to fetch the journey map from
the resources as a stream, rather than a file.

When importing the process-journey-event lambda into a different project
(the local-running gradle project), loading as a file breaks. The file
path is a jar file path to the process-journey-event jar and for unknown
reasons the file can't be found withing the generated jar.

Switching to use streams works just fine.