PYIC-3609: Allow all options to be used
Updates the html and scripts to allow multiple CRIs to be disabled and
multiple flags to be activated.

Updates the way the CRI and flag options are gathered by
recursively checking the events. Setting a disabled or flag check in
the journey map allows for events to be defined underneath it. These may
have options not used at the top level else where.

Update to consider these sub events when resolving the target states.
Previously we were only resolving the top level.