BAU: Add debugging for core-front
This overrides core-front's docker images command to allow a debugger to
be attached. It also exposes port 5001 which is routed to the default
node debugger port.

This also sets the `CDN_DOMAIN` for core-front to null. This was
previously being interpreted as `undefined` by node which caused some
static files to fail to load, causing some noise in the logs.