BAU: Add tests for successful VC fix
A change was recently merged around how we check the success of VCs when
checking for the fail-with-no-ci scenario. We shouldn't include the CIs
in a VC in that check. It was merged to fix an incident.

This is an update to the tests to ensure we prevent this happening
again.