PYIC-3658: Update IPV_F2F_CRI_VC_CONSUMED audit event
The way we were adding name parts to the restricted section of the audit
event was out of line with the rest of the programme. This changes it
from `restricted.nameParts` to `restricted.name.nameParts`.

Worth noting that this is the only place that we include restricted data
in an audit event. We don't send this with the normal VC retrieval
method.

Also, we include the `docExpiryDate` attribute in the restricted data.
If the VC contains neither passport or driving license data, this
attribute is included but set to null. We should potentially just
exclude this attribute if its null, however I'm unsure how these events
are used so am favouring leaving it in. In reality we shouldn't be
getting VCs that don't include passport or DL data so this may be a non
issue.