PYIC-3658: Wrap `nameParts` in an array
Looking at other audit events that include the name and name part
attributes in the restricted section of the audit event
(DCMAW_CRI_VC_ISSUED and IPV_KBV_CRI_REQUEST_SENT for example), the name
parts are objects within an outer array. If this work is to bring how we
are sending this data in line with other audit events, we should be
consistant.