BAU: Update local-running Dockerfiles
The dockerfiles that were being used have had Dynatrace added to them.
As part of that a remote resource is fetched from a container registry
that is behind auth. Which makes it hard to build locally.

New dockerfiles were created that were the same as the old ones but
didn't include dynatrace. We can use those instead.