PYIC-3829: Return exit code for contra indicators
We want to return an exit code attribute in the user identity fetched by
the orchestrator.

This change maps the CI codes a user has to the appropriate exit codes
and adds them to the user identity. This is hidden behind a feature
flag.

In the case that a user has achieved a P2, we only want to add exit
codes for specific CIs. These are defined in config to avoid leaking
sensitive information in this public repo.

In the case that a user has achieved a P0 due to breaching the CI
threshold, we want to map all CI codes to their exit codes. This is
looked up in a mapping table defined in config. This includes mitigated
CIs. This is because a mitigated CI will still have a residual score,
and so will be contributing to the threshold breach.

The case that a user has achieved a P0 but not due to a CI threshold
breach will be handled in another ticket.