PYIC-3869: Extend featureSets to backchannel lambdas
Feature sets are set in the config service by a value passed through in
the request. This value comes from the users frontend session.

Lambdas that are called directly by clients don’t have access to the
value from the frontend session, meaning that the feature sets can’t be
used.

We can store the feature set in the users backend session at the last
appropriate moment - in the build-client-oauth-response handler. This
makes it available in those lambdas not called as part of a users
request.