PYIC-3924: Return exit code for non breaching failures
This extends the exit code functionality to return a specified exit code
in the scenario that a user has failed to prove their identity (`P0`),
but is not breaching the CI threshold limit. This would usually indicate
that the user has chosen not to finish the journey. Possible reasons
being they've go no documents, or don't want to/can't use the app or F2F
journeys.

This is distinct from the scenario where a user abandons their journey
entirely, which is not in scope.