PYIC-3947: Remove EXIT_CODES feature flag
The flag is set to true all the way out to production. We can now remove
the flag itself.

This also removes the `NON_NULL` annotation from the exit code field in
the UserIdentity class. It was only there to prevent the field being
returned to orchestrator before we were ready.