BAU: Add orphan states option to journey map render
This adds a new option to the rendering of the journey map to show
orphan states - states that have no other state targeting them.

We only have two states that we expect to be in this state - the `INITIAL_IPV_JOURNEY`
state and the `CORE_SESSION_TIMEOUT` state. Both of these states are set
directly in the core back code, rather than being transitioned to in the
usual flow.

The `INITIAL_IPV_JOURNEY` state has been handled, but the
`CORE_SESSION_TIMEOUT` state has been left in. Perhaps we should also
exclude it...