PYIC-3960: Remove f2fCimit feature flag
This flag was just used in the jourey map, not in the java code. The
flag is set to true all the way out to prod and we've seen users
completing the journey successfully. We can now remove the flag.

This leaves the disabled check on the f2f CRI in case we do want to
disable it for any reason (like we already have).