PYIC-3991: Fix exit code breaching threshold logic
When we check a users CI scores against the threshold we only fail them
if the score is *greater* than the threshold. Equal to is ok.

When calculating the exit code we were deciding the scores were
breaching if they were greater than *or* equal to. Which is wrong.

This pulls the logic for determining if CI scores are breaching to a
shared place. This will ensure that we're always determining it the same
way.