BAU: Fix some code smells
We're currently not getting the callback from Sonar in our PRs. This is
a known issue and is waiting for the GitHub org migration to be 100%
complete.

This fixes up a few code smells that have slipped through the net.

See the report here: https://sonarcloud.io/project/issues?resolved=false&types=CODE_SMELL&sinceLeakPeriod=true&id=ipv-core-back