PYIC-4056: Make VerifiableCredentialException checked
This exception was extending a RuntimeException, rather than just
Exception. This meant that it wasn't a checked exception. There was no
real reason for this - one piece of logic needed to be rewritten as a
simple for loop.
Making it a checked exception means that we aren't going to miss
handling it.