PYIC-4056: Fix CheckExistingIdentityHandlerTest
A recent change was made to the structure of the UserIdentityService.
Specifically we only call the `areVcsCorrelated` method to check
correlation. This means the mocking we had set up was a bit wrong - it
was mocking the internal methods called by that method. This was messing
up return value from the mock and breaking the test.

Not entirely sure why these tests passed in the first place.