BAU: Run gradle tasks in parallel
Our gradle project consists of lots of separate sub-projects, without
too many dependencies on each other. This makes it a good candidate to
run separate sub-project tasks in parallel.

Currently, on my machine, running `gradlew clean build` takes 2m12s.
Applying the parallel config reduces that to 59s.

We'd need to keep an eye on how this affects build in GHA for the
pipeline, but hopefully it should be fine.