PYIC-4056: Add method to ConfigService for active connection API key
A backend only CRI will only ever need to use the active connection
configuration for that CRI. A typical OAuth based CRI needs to use the
same connection for all interactions across the users session. This is
not the case for back end only, which just calls the CRI once.

This adds a new method to get the API key from secrets manager for the
active connection for a CRI.

It also tidies some private methods to beneath the public.