PYIC-4056: Add ConfigService method for back end cri config
The config required for a back end only CRI is a lot simpler than the
usual config for a CRI. This is because it doesn't use OAuth.

This adds a class and method to represent the required config. This
should make working with the config object simpler and less prone to
errors.