PYIC-4056: Pass lambdaInput in process request directly
This changes the way the lambdaInput, defined in the journey map for a
process step response type, is passed to the lambda via the process
request.

It passes the object through as a whole, rather than pulling out
individual elements. This makes it a lot more flexible. Otherwise we'd
need to add new fields to the ProcessRequest class every time we wanted
to pass some data through from the map.