PYIC-4056: Add logic for calling TICF CRI
This fleshes out the handler, which orchestrates various services, and
adds a new service for talking to the TICF CRI.

This still needs plumbing in to the journey map, so this commit on its
own won't make any functional changes.

When adding to the journey map it will need to have a `lambdaInput`
block defined, to pass through the `journeyType` value. This is due to
a requirement that reuse journeys don't send any VCs to the TICF CRI.