PYIC-4200: Serve well known file from S3
This updates core's template to serve the well-known jwks endpoint from
an S3 bucket instead of from the stubbed mock data.

This is dependant on that file existing. The mechanism for that hasn't
been decided yet and this should not be merged until it is, or bad
things™️  will happen.