INCIDENT-599: Stop parsing reprove identity claim
A change has been made in Orch that has meant that the value coming
through is no longer a string - possibly null.

This is causing the `getStringClaim()` method to throw. This has caused
a P1 outage.

The change on orch side is being reverted - this change is a backup just
in case to return functionality.