BAU: Turn off xray logging in local running
When using the local running setup X-RAY tracing doesn't work, due to
not passing around the correct HTTP headers. We don't really need it to
work when running in this environment anyway.

There is no global way to "switch-off" x-ray without removing all the
code and the library which isn't practical.

This results with a LOT of annoying log messages from core-back when
running locally.

We can just turn off the logger for the x-ray class which is responsible
for emitting these messages. Unfortunately x-ray uses log4j 1.2, even
though it's been deprecated since 2015. As we're using log4j 2.x, x-rays
loggers were falling back to using the default java logger and as such
our log4j config file was having no effect on the annoying log messages.

The solution is to include the log4j bridging lib, which allows 3rd part
libs using the 1.2 API to work with a 2.x set up.