BAU: Remove DDB TTL from UserIssuedCredentialsV2Table
This removes the config from the SAM template that enables the DynamoDB
TTL feature for the UserIssuedCredentialsV2Table table. This will have
the effect of no more VCs being removed automatically from the table
once the timestamp in the ttl attribute of the record has been reached.

It also removes the writing of the ttl timestamp for any future VCs
stored in the table as this is no longer required.