PYIC-4331: Track VCs to submit to TICF
We want to simplify the journey map when invoking the call-ticf-cri
lambda by not having to supply a journey type. The journey type was
being used to determine if we should send all the users VC (an IPV
journey), or none (a reuse journey).

What we're actually interested in sending to the TICF VC is the VC
gathered during the users current session. The previous behaviour was
just a proxy for that.

This change keeps track of the VC the user has received in the current
session, and then only sends those to the TICF CRI when appropriate.

I had originally planned to just store the JTI from the VC in the users
session, and use that to filter the users VCs from the store. However
not all CRIs currently provide a JTI claim (even thought they should),
so have fallen back to storing the entire serialized VC in the session.

I've maintained the filtering behaviour - the VC in the users session
are compared against the VC from the VC store before sending. We could
in theory just use the VC in the store directly and skip the check of
the VC in the store - this would save a round trip to the DB. It feels
safer however to do the check, although I don't have a good example of
why this is beneficial.

One consequence of this change is that on a reuse journey, where the
user decides to reset their identity (an upcoming feature) we could send
TICF their own VC. Hopefully this is ok.