PYIC-4217: Fix NPE when name parts are null
The code we had to create a full name from a VC's name parts was
throwing NPEs. This would happen if one of the name parts was null. We
were calling equals on it when trying to filter out values that were
equal to an empty string.

We've determined that it is not currently acceptable for a VC to have a
missing name part, so we should be catching and logging this scenario in
advance of getting to the stage of throwing this NPE.

And we were infact already attemting to do this when retrieving the name
parts from the VC. We were only throwing though if ALL name parts were
missing. This updates that check to catch if ANY of the name parts are
missing. It also refines that check to check for null, empty string OR
pure whitespace.

This change also removes the code that was originally throwing the NPE
as it was unnecessary. Even in the scenario where a name was set to an
empty string, there was no functional difference if it was filtered out
or not, as the empty string would just be joined with other strings and
dissapear.

This also updates the tests to be more flexible. Rather than relying on
hardcoded VCs, which can be difficult to understand what they're doing,
it allows VCs to be constructed with specific names and DOBs.