BAU: Supress warning about singleton pattern
Sonar flags that the singleton pattern should be used with care.

The singleton pattern was implemented to allow for easier testing for
the contract tests. It was difficult to mock the values (required for
the tests) when it was just being called as a static method. The pattern
allows the entire class to be mocked correctly.