BAU: Handle InterruptedException
The http client can potentially throw an InterruptedException. This
would happen if another thread tried to interrupt the current one while
it was in a waiting or sleeping state. This won't happen in Lambda as
it's single threaded.

Sonar flags this a bug howvever suggests we need to call interrupt().
For us, this means that the http request has failed and we should
continue processing the current thread as though the TICF CRI was
unavailable.