PYIC-4258: Handle inheritedIdentityJwt Failures
If there is an issue with the received inherited identity jwt - badly
formatted, or doesn't validate - we don't need to just return a technical
error. By the point we're throwing an error we'll have already validated
the clients redirect URI. That means we can close the OAuth loop and
send the user back.

This also adds domain objects to parse the claims/userinfo json blob
into. It looks like we won't be receiving the inhertied identity jwt on
the top level, but rather with the other requested claims (which we
currently ignore).