PYIC-4258: Fix equals methods across CRI hierarchy
This gives each level of the CRI config a fully working equals method.
They can be used in our tests to simplify things a bit.

It moves the getter for the parsed keys to a method that doesn't clash
with the default getter - this allows the equals methods to work.