PYIC-4258: Allow no inherited identity claim
Most of the time we won't be receiving an inherited identity jwt, so we
need to allow this to happen once the feature flag is removed.

This checks the claims claim and the userinfo object and the inherited
identity claim. If any of them are null we return.

In reality we should always be getting a claims claim and a userinfo
claim, as they're currently in the JAR from orch, but should still do
the null checks when diving into this object. Perhaps we should be
throwing though if they are null.