PYIC-4258: Correctly convert claims claim
This change ensure that we're correctly converting the values in the
claims claim. Previously we were just casting them. This worked as the
test values were being set directly in the claimsset object and not
undergoing any deserialization/serialization.

This updates the tests to ensure that the value are the same as those
we'll receive over the wire. They are actually presented to us as
JSONObjects, which we can convert into the domain objects we can work
with.