BAU: Fix message on InvalidCriCallbackRequestException
I recently change the parent of the InvalidCriCallbackRequestException
to a raw Exception. I failed to call super() with an error message. This
meant that when the error was being thrown, where we were previously
logging the message from the exception, now we're just logging null.

This should reinstate the expected behaviour.