PYIC-4515: Evaluate vtr and include operational profiles
This updates the check-existing-identity lambda to take account of the
vtr value received in the JAR request from orch. It will now check if
the user has met any of the vtr received, in strength order, and return
a reuse journey if the have.

It creates some enums for the vot we will support, and maps them to the
profiles that are currently supported for a user to meet that vot.

For a GPG45 vot (currently only P2), this is done in the usual way. For
an HMRC vot this is done by checking if the user has an inherited
identity VC that has the correct level of vot set on it.