PYIC-4515: Refactor Vot enum
This updates the Vot emum to set null values rather than empty lists for
the profile types that aren't supported. This allows us to determine if
a Vot is of Gpg45 type of not just by checking that the list of gpg45
profiles is non null.

It would have been good to have two separate constructors that would do
the setting of null for us, rather than having to expliclty pass the
null to the constructor. It would enforce that there is only one list
populated. But after type erasure the constructors would have the same
signature and this isn't valid. There are ways around this, but they
could be a bit horrible.

This also changes `valid` for `supported`.