BAU: Fix local running orch stub
The orch stub has changed to be multi env and now needs a value for a
default jar encryption key. In the local env you'll only need the
default, so we can just update the current env var and not add the other
environments.

This also adds a signing key for the inherited identity jwts.