PYIC-4548: Simplify attainedVot logic
This simplifies the return values of the methods used to check if a user
has attained a gpg45 VOT or an operational VOT. We were returning a
VotProfilePair and only using the pair for some logging.

This removes that record, and just returns a boolean for if the
requested VOT being checked has been met or not.