BAU: Only fetch ci-config if required when validating VC
When validating VCs we validate that any CI codes contained within are
valid. If the VC doesn't have any CIs we don't need to fetch the config.

This has become an issue due to the validation of inherited identity VCs
which won't contain CIs. Without this change the initialise-ipv-session
lambda would need to have read access to the ci-config secret, even
though it doesn't actually require that access.