PYIC-4601: Route users depending on reuse type
This updates the check-existing-identity lambda to send users on the
correct journey, depending on if they are reusing an identity and if
they are in the process of migration or not.

Users reusing an operational profile that have already migrated (in a
previous session) should see the reuse page but with some extra context.
Those that are migrating in the current session should just go straigh
back to the service.

This has required some extra CRI_TICF states to achieve, due to the new
follow on routes that we didn't support before. Namely ending
immediately and showing the reuse page with context.

This has added a new CRI_TICF parent state to the joureny map to dry up
some of the existing states.