PYIC-4975: Allow operational profiles with pending F2F
If a user has an operational profile they should be able to use it, even
when they have a pending F2F request.

This slightly reorders the logic in the check-existing-identity lambda
to allow this. We only return a pending f2f response once we've checked
that they're breaching CIs, or if they haven't matched any profiles
(gpg45 or operational).