PYIC-4982: Defer setting of signing key ID
The process-cri-callback lambda uses the CriApiService to generated
token requests for CRIs.

These token requests are signed, and we use a custom KMS signer to do
this. The signer needs a KMS key Id configuring on it.

We were doing this in the initialisation phase of the lambda. Moving to
snapstart would meant that value was permanently baked in to the lambda.
This would mean we couldn't switch signing keys easily if required.

This change defers the setting of the key id to the runtime. We use a
KMS signer when running in a live enviroment, and a normal ES245 signer
in the tests, which is why we need to check if the signer is of a
particular type before trying to set the KMS key ID on it.