PYIC-4980: Disable snapstart in build temporarily
Snapstart needs provisioned concurrency to be disabled in order to
activate. And provisioned concurrency needs snapstart to not be active
to be disabled.

This temporarliy turns snapstart off in build, so we can stop the
provisioned concurrency. Once that has happened we can turn snapstart
back on.

Once build is running, I'll be experimenting with my dev env to try and
find a way to do this in prod without having to run without provisioned
concurrency or snapstart.