PYIC-5136: Handle cimit VC with documents array
The cimit VC will contain CIs with an array of document IDs, rather than
just one.

This adds a new domain model to handle the new format, and a custom
deserializer to ensure we're deserializing using the correct version.

Currently we're not using the document identifier here, so there are no
changes to how we're handling the deserialized contra indicators.