BAU: Fix local running
The local running setup was failing due to how we were constucting the
input to the criCallBack handler - if no feature sets were active we'd
get an NPE.

This updates to invoke the handleRequest method on that handler, rather
than calling one of the internal methods. This allows the existing
helpers to parse the feture sets and do the right thing with it.

Also some refactoring to change a big if statment to a case statement.