BAU: Add a test to local-running
Without any tests jacoco won't produce a test report. Which means that
sonar will determine that new code isn't tested. Which to be fair it
isn't. But PRs fail.
The local-running config is only for running core locally as a dev tool,
so the lack of tests shouln't be too concerning.