PYIC-5343: Don't override status code from process journey event
The ErrorStepResponse class has a statusCode field. This class
represents the json that is returned from the process-journey-event
lambda when a users transitions to a state of type "error".

The vtl code in the openAPI spec was using the status code in the json
to override the status code of the response to the frontend.

The `ERROR_PAGE` state sets the status code to 500 (this is actually the
only state that uses the ErrorStepResponse class). When core-front
received the response from core-back, with the status code overriden to
500, axios was intercepting it and throwing an error. This caused the
wrong page to be displayed - the default error page rather than a
recoverable one.

This change prevents the overriding from happening, so axios will stay
out of the way. Instead the request will have a 200 status code. This is
actually correct, as the backend has fulfilled its role quite happily.

With just this change the user will see an error page, but will get a
200. A follow up PR will get the frontend to honour the status code
returned to it.