BAU: Allow multiple same CI codes
Core was blowing up when a user had more than one CI with the same code.
This is likely to happen with the new mitigation work.

Internally we were converting the list of CI received from CiMit into a
map, with the CI code as the key. This doesn't work for multiple same CI
codes.

This change updates the internal structure to a list, and fixes up all
the methods that previosuly operated on the map to operate on the list
instead.