PYIC-5132: Define dependencies in libs.versions.toml
This creates a "version catalogue" in gradle - a central place to
declarte dependencies and their versions which can be shared by
sub-projects. For the gradle docs see the link below:
https://docs.gradle.org/current/userguide/platforms.html

We were previously declaring versions in an extension in the root of the
project to do a similar job. Whilst this worked, it meant that
dependabot was unable to raise PRs to bump versions. The dependabot docs
suggest that it can work with a libs.versions.toml:
https://docs.github.com/en/code-security/dependabot/dependabot-version-updates/configuration-options-for-the-dependabot.yml-file#github-actions

Makes sense to use the proper tool for the job.