PYIC-5132: Commonise aws sdk version
This commonises the version for packages from aws published under
`software.amazon.awssdk`.

There was already a version with that name, so it's been moved to the
more specific `awsJavaSdk` for pacakages published under
`com.amazonaws:aws-java-sdk-*`

I'm mostly interested in how dependabot handles this. It's raised PR's
to bump the version number - I want to see if it recognised a shared
version...