BAU: Upgrade nimbusds to 11.10.1
This was a suggestion from dependabot, but there were a few breaking
changes to deal with.

The most notable is the removal of the inbuilt JSON lib that we were
using in places. This change replaces it with Jackson. It also allows us
to get rid of gson in a few places.

There are a few other small changes, such as the JWEDecrypter#decrypter
method having an extended signature, but nothing very big.